create mars.py
Adding the tutorial code to the repo in its current state. I'm having a lot of difficulty getting the code into GitBook in a state that you can easily copy/paste it into Notepad ++ without losing the formatting (tabs/spaces). Thought we might have better luck pointing people to a link to this code in the repo and copy/pasting that way? Adding it to confirm.